Handle Mac coordinate flip in CairoQuartzSurface
These changes are associated with commit 549d208 in the Tk package.

This simplifies things greatly, because Tk widgets also get flipped with the old way of doing things.  We now flip the y axis when making a Cairo Quartz surface, then flip back immediately afterwards.  The initial graph in Winston is upside down, but corrects itself on a resize.  I think this remaining bug is related to the order of callbacks, which I haven't quite figured out yet.